Rearrange ProfileScreen code
