+++ courselynx/app/settings/profile/ProfileScreen.tsx
@@ -1,7 +1,8 @@
 import { useState } from "react";
-import { View, Text, Image, TouchableOpacity, StyleSheet, ScrollView } from "react-native";
+import { View, Text, Image, TouchableOpacity, ScrollView } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import { InputField } from "@/components/InputField";
+import styles from "./ProfileStyles";
 
 type ProfileProps = {
   avatar: string;
@@ -72,7 +73,6 @@ const Self = ({ userdata }: { userdata: ProfileProps }) => {
 }
 
 const Public = ({ userdata }: { userdata: ProfileProps }) => {
-  // Sample data for groups
   const socials = [
     { name: "LinkedIn", icon: "logo-linkedin", color: "#0A66C2", username: "@abcxyz" },
     { name: "Discord", icon: "logo-discord", color: "#5865F2", username: "N/A" },
@@ -84,9 +84,13 @@ const Public = ({ userdata }: { userdata: ProfileProps }) => {
     { name: "Dance Club", color: "#D39AD7" },
   ];
 
+  const messageUser = () => {
+    console.log("Message");
+  }
+
   return (
     <>
-      <ScrollView contentContainerStyle={styles.container} automaticallyAdjustKeyboardInsets={true}>
+      <ScrollView contentContainerStyle={[styles.container, styles.lightBackgroundColor]} automaticallyAdjustKeyboardInsets={true}>
         {/* Profile Header */}
         <View style={[styles.profileContainer, styles.horizontalContainer]}>
           <Image source={{ uri: userdata.avatar }} style={styles.avatar} />
@@ -129,116 +133,13 @@ const Public = ({ userdata }: { userdata: ProfileProps }) => {
 
       {/* Message Button */}
       <View style={[styles.fixedButtonContainer, styles.lightBackgroundColor]}>
-        <TouchableOpacity style={styles.button}>
+        <TouchableOpacity style={styles.button} onPress={messageUser}>
           <Text style={styles.buttonText}>Message</Text>
         </TouchableOpacity>
       </View>
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flexGrow: 1,
-    padding: 20,
-    paddingTop: 10,
-  },
-  profileContainer: {
-    alignItems: "center",
-    justifyContent: "center",
-    marginBottom: 10,
-  },
-  horizontalContainer: {
-    flexDirection: "row",
-    flexWrap: "wrap",
-    gap: "4%",
-  },
-  lightTextColor: {
-    color: "#111",
-  },
-  lightBackgroundColor: {
-    backgroundColor: "white",
-  },
-  label: {
-    fontSize: 14,
-    fontWeight: "bold",
-  },
-  text: {
-    fontSize: 12,
-  },
-  avatar: {
-    width: 150,
-    height: 150,
-    borderRadius: 75,
-  },
-  name: {
-    fontSize: 20,
-    fontWeight: "bold",
-    marginBottom: 5,
-  },
-  details: {
-    fontSize: 15,
-    marginBottom: 5,
-  },
-  bio: {
-    textAlign: "center",
-    fontSize: 14,
-    padding: 10,
-  },
-  sectionTitle: {
-    fontSize: 16,
-    fontWeight: "bold",
-    marginTop: 20,
-    marginBottom: 8,
-  },
-  item: {
-    flexDirection: "row",
-    gap: 10,
-    paddingVertical: 10,
-    alignItems: "center",
-  },
-  groupColor: {
-    width: 40,
-    height: 40,
-    borderRadius: 10,
-  },
-  form: {
-    marginBottom: 10,
-  },
-  editButton: {
-    position: "absolute",
-    bottom: 5,
-    left: "50%",
-    transform: [{ translateX: 45 }],
-    backgroundColor: "#3b82f6",
-    width: 30,
-    height: 30,
-    borderRadius: 15,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  fixedButtonContainer: {
-    position: "absolute",
-    bottom: 0,
-    left: 0,
-    right: 0,
-    padding: 20,
-    paddingTop: 10,
-    borderTopWidth: 1,
-    borderColor: "#ddd",
-  },
-  button: {
-    backgroundColor: "#3b82f6",
-    paddingVertical: 14,
-    borderRadius: 8,
-    alignItems: "center",
-  },
-  buttonText: {
-    color: "white",
-    fontSize: 16,
-    fontWeight: "bold",
-  },
-});
 
 const ProfilePage = { Self, Public };
 export default ProfilePage;
